requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
Deprecated==1.2.13
idna==3.3
pycparser==2.20
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wrapt==1.13.2
